Fix/awt 42 fix string util and goToProject (#19)
* Restore and fixed StringUtilTest

* Minor changes

* AUpdated Dashboard to fix go to project. Updated addstory button findby

* Updated dashboard gotoproject. reverted story

* minor change

* added missing java doc, added missing end new line.

* Removed log file.

* Removed test file.

+++ src/main/java/org/fundacionjala/core/ui/AbstractPage.java
@@ -15,7 +15,7 @@ public abstract class AbstractPage {
     protected WebDriver driver;
     protected WebDriverWait wait;
     protected WebDriverAction action;
-    public static final int WAIT_TIME = 60;
+    public static final long WAIT_TIME = 60;
 
     /**
      * Constructor of class.
